Reverted code move in AsyncTask & doc'd.
After chatting with a friendly android dev about the location of
post network call functional work, the point was raised that:

* `doInBackgroud()` is completely non-UI thread.
* `onPostExecute()` is UI thread.

Therefore, `onPostExecute()` should contain the minimal ammount of
functional work as possible since it can lead to noticeable lag on
the UI when running at 60fps.

Documented for future me to remember.