Made Service Generator generic.
I want to move onto doing some Oauth learning but for that I want
to hit a different API. Seemed silly to do copy/pasta just to use
a different url, so changed the Service Generator to be generic.

Put base url on the client, to keep everything in one place. Might
change it to be auto read from the client.