Implement first stage of Oauth.
Lets implement the first stage of Oauth where we get the User to
log into Fitbit, and we present them with what access/expiry we
want and they can choose allow/deny.

Post User making a choice we currently display the returned code
or error message.

This uses an intent-filter to capture a unique schema that the
browser doesn't understand, to send the redirect_uri back to
the application.