Get access_token if we already have access.
So, I haven't saved the access_token from the initial Oauth pass,
which means that on subsequent presses of the `Fitbit Login` button
we hit the Browser and comeback straight away with a Uri containing
`user_id` and `access_token` (among other things).

Since this uri is a fragment (instead of a query), there is some
manual string munging that is needed to pull out what we need.

Still not making authenticated requests, but at least I'm now
pulling out the `user_id` and `access_token` for later.