Fitbit Activity initial boiler plate.
First part of the boiler plate work to add the Fitbit API. Starting
with the first Activity call, by adding a model and an API call.

Will plumb in later on with an appropriate async retrofit2 call or
synchronous Retrofit2 + AsyncTask bit of code. Either way, I'll
have to pass in the `user_id` & `access_token` from the prior
Fitbit Oauth step.